explicitly cast v-for index as number
When building for production (and only production) ts emits an error:

```
ERROR in /Users/johnsojr/code/Camino/resources/camino-creator/components/Stage/stages/DeepDives.vue.ts
112:53-58
[tsl] ERROR in /Users/johnsojr/code/Camino/resources/camino-creator/components/Stage/stages/DeepDives.vue.ts(112,54)
      TS2345: Argument of type 'string | number | symbol' is not assignable to parameter of type 'number'.
  Type 'string' is not assignable to type 'number'.
```

This is unhelpful for several reasons:
1. the error references line 112, but the .vue file has less than 100 lines, so presumably this is the line number after some transpiling step.
2. linting errors show up in VSCode (or WebStorm for that matter)
3. Everything builds fine for dev. The error only occurs when building for production.

Ultimately, I traced the error to the `index` value in the v-for statement.

It seems that TS has trouble typing the index value as a `number` in the DeepDives component (again, only when building for production).

One hypothesis is that `v-for` can be used on objects AND arrays, so the index could be a string or a symbol (if object) or number (if array). And maybe some transpiling step for production throws away the necessary context to distinguish whether this is an object or an array.

Explicitly declaring `index as number` when used in the v-for loop makes the issue go away.

I suspect the issue may crop up again with indices when building for production, so it might be best to always explicitly declare the index type with v-for loops?