keep map from rerendering on each stage change
Previously, whenever a stage would update within a tour stop, it would trigger the map component to rerender (even if it wasn't a navigation stage update). To users, the presses would feel laggy, and the map lines would flicker as the map would redraw.

The current hypothesis is that a stage change is triggering the whole map redraw because finding a valued target point is reading each stop and each stage (we need to loop through all the stages to check if they're a nav stage or not). Because these points are computed, once tour stage change cascades down.

The workaround is to get the tour and all the stop points and routes only once at setup time. Thus, in this commit, we:
- remove reactivity from the `tour`, making it a plain object.
- and calculate mappedStops as a plain array, rather than a computed ref.

A better fix may be to restructure the Creator Store. Instead of each tour being one giant object, as returned from the API, maybe flatten it a bit so that stops and stages have their own lookups. (perhaps even structuring by stage type?)