improve stage type, quizStore refactor
Previously it was possible for TS to miss stage errors on properties that don't really exist. This tightens up the Stage so that it has to be one of the known stage types, and creates a CoreStage class that's extended.

chaning the quiz attempt state wasn't working. This was traced to an issue in quizStore where I used the spread operator when returning data from a getter like `{...quiz, ..quizAttempt}` which killed reactivity of the quizAttempt. The object status returned from the getter would be different than the object status in the store.

Also: the quiz store could possible have non-initialized valued for a given quiz. This adds an `init()` method to initialize the quiz store