track when stops are done
When a user completes all quizzes we don't want to close the modal right away. Instead, we want them to click "Continue" so that they have a chance to review their answers.

This adds a list in quizStore to track which stops are truly done (e.g. all quizzes completed AND the user hits continue).