revert bootstrap to 5.1.3
In 5.2.0, boostrap changed to CSS variables, which altered how some things render. For example, on the login page the button hover state is transparent, making it unreadable. This is maybe due to a bug with an unset CSS variable?

Whatever the case, this commit rolls back to 5.1.3 for this release, since there's nothing in 5.2.0 that compels an update.